Fix prototypes, adding const

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: adda-sil <adda-sil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/10 19:38:33 by adda-sil          #+#    #+#             */
-/*   Updated: 2019/10/21 17:07:11 by adda-sil         ###   ########.fr       */
+/*   Updated: 2019/10/23 18:43:06 by adda-sil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include <string.h>
+
+size_t	ft_strlen(const char *str)
 {
 	const char *ptr;
